+++ source-code/clj/server_extensions/storage/engine.clj
@@ -46,17 +46,21 @@
 
 (defn attach-item!
   ; WARNING! NON-PUBLIC! DO NOT USE!
-  [{:keys [request]} directory-id {:media/keys [id] :as media-item}]
-  (letfn [(prototype-f [document] (mongo-db/updated-document-prototype request :media document))
-          (attach-f    [document] (update document :media/items vector/conj-item {:media/id id}))]
-         (mongo-db/apply-document! "storage" directory-id attach-f {:prototype-f prototype-f})))
+  [_ directory-id {:media/keys [id] :as media-item}]
+  (letfn [(attach-f [document] (update document :media/items vector/conj-item {:media/id id}))]
+         (mongo-db/apply-document! "storage" directory-id attach-f)))
 
 (defn detach-item!
   ; WARNING! NON-PUBLIC! DO NOT USE!
-  [{:keys [request]} directory-id {:media/keys [id] :as media-item}]
-  (letfn [(prototype-f [document] (mongo-db/updated-document-prototype request :media document))
-          (detach-f    [document] (update document :media/items vector/remove-item {:media/id id}))]
-         (mongo-db/apply-document! "storage" directory-id detach-f {:prototype-f prototype-f})))
+  [_ directory-id {:media/keys [id] :as media-item}]
+  (letfn [(detach-f [document] (update document :media/items vector/remove-item {:media/id id}))]
+         (mongo-db/apply-document! "storage" directory-id detach-f)))
+
+(defn item-attached?
+  ; WARNING! NON-PUBLIC! DO NOT USE!
+  [_ directory-id {:media/keys [id] :as media-item}]
+  (boolean (if-let [{:media/keys [items]} (mongo-db/get-document-by-id "storage" directory-id)]
+                   (vector/contains-item? items {:media/id id}))))
 
 
 
@@ -68,7 +72,8 @@
   ;
   ; @param (map) env
   ; @param (namespaced map) media-item
-  ;  {:media/filesize (B)
+  ;  {:media/content-size (B)
+  ;   :media/filesize (B)
   ;   :media/path (namespaced maps in vector)}
   ; @param (function)(opt) operation
   ;  -, +
@@ -77,16 +82,17 @@
   ([env media-item]
    (update-path-directories! env media-item nil))
 
-  ([{:keys [request]} {:media/keys [filesize path]} operation]
-   ; Mappák és fájlok létrehozásakor/feltöltésekor/törlésekor szükséges a tartalmazó (felmenő) mappák
-   ; adatait aktualizálni:
+  ([{:keys [request]} {:media/keys [content-size filesize path]} operation]
+   ; Mappák és fájlok létrehozásakor/feltöltésekor/törlésekor/duplikálásakor szükséges
+   ;  a tartalmazó (felmenő) mappák adatait aktualizálni:
    ; - Utolsó módosítás dátuma, és a felhasználó azonosítója {:media/modified-at ... :media/modified-by ...}
-   ; - Tartalom mérete {:media/content-size ...}
+   ; - Tartalom méretének {:media/content-size ...} aktualizálása
    (letfn [(prototype-f [document] (mongo-db/updated-document-prototype request :media document))
-           (update-f    [document] (if operation (update document :media/content-size operation filesize)
-                                                 (return document)))
+           (update-f    [document] (cond-> document content-size (update :media/content-size operation content-size)
+                                                    filesize     (update :media/content-size operation filesize)))
            (f [path] (when-let [{:media/keys [id]} (last path)]
-                               (mongo-db/apply-document! "storage" id update-f {:prototype-f prototype-f})
+                               (if operation (mongo-db/apply-document! "storage" id update-f {:prototype-f prototype-f})
+                                             (mongo-db/apply-document! "storage" id return   {:prototype-f prototype-f}))
                                (-> path vector/pop-last-item f)))]
           (f path))))