+++ source-code/clj/server_plugins/item_editor/sample.clj
@@ -1,12 +1,11 @@
 
 (ns server-plugins.item-editor.sample
-    (:require [mid-fruits.candy     :refer [param return]]
-              [mid-fruits.string    :as string]
-              [mid-fruits.validator :as validator]
-              [mongo-db.api         :as mongo-db]
-              [pathom.api           :as pathom]
-              [x.server-core.api    :as a]
-              [x.server-db.api      :as db]
+    (:require [mid-fruits.candy  :refer [param return]]
+              [mid-fruits.string :as string]
+              [mongo-db.api      :as mongo-db]
+              [pathom.api        :as pathom]
+              [x.server-core.api :as a]
+              [x.server-db.api   :as db]
               [com.wsscode.pathom3.connect.operation :as pathom.co :refer [defresolver defmutation]]
               [server-plugins.item-editor.api        :as item-editor]))
 
@@ -24,7 +23,7 @@
   (let [item-id (pathom/env->param env :item-id)]
        (if-let [document (mongo-db/get-document-by-id "my-collection" item-id)]
                ; XXX#6074
-               (validator/validate-data document))))
+               (pathom/validate-data document))))
 
 (defresolver get-my-type-item
              ; @param (map) env