+++ source-code/cljs/app_extensions/storage/media_browser/dialogs.cljs
@@ -48,7 +48,7 @@
                          :on-click [:storage.media-browser/rename-directory! directory-item]}]
        [elements/button ::delete-directory-button
                         {:preset :warning-button :icon :delete_outline :indent :left :label :delete!
-                         :on-click [:storage.media-browser/delete-directory! directory-item]}]])
+                         :on-click [:storage.media-browser/delete-item! directory-item]}]])
 
 
 
@@ -80,7 +80,8 @@
                         {:preset :default-button :icon :edit :indent :left :label :rename!
                          :on-click [:storage.media-browser/rename-file! file-item]}]
        [elements/button ::delete-file-button
-                        {:preset :warning-button :icon :delete_outline :indent :left :label :delete!}]])
+                        {:preset :warning-button :icon :delete_outline :indent :left :label :delete!
+                         :on-click [:storage.media-browser/delete-item! file-item]}]])